♻️ Move Fingers to $lib/touch/fingers to make space for future Gestures class
